Fix token counting: count function_call args, calibrate chars-per-token
A live chat reached 281,152 real tokens while the estimator reported
183,684 — a 1.53x undercount that let the conversation blow past every
budget until the model confabulated. Measured decomposition on that
419-message thread:

  content (real tokenizer) : 207,492
  function_call arguments  :  51,179  <-- NEVER COUNTED
  system prompt            :   3,220
  tool schemas (52 tools)  :   8,120
  chat-template framing    :  ~11,000
  ------------------------------------
  server actually saw      : 281,152

Two real defects:
1. estimate_message_tokens read only msg.content — a message carrying a
   function_call was charged a flat 10 tokens, so 131 tool calls worth
   of arguments (project_write_file passes whole file contents) were
   invisible. Now counts function_call name+arguments and
   reasoning_content.
2. chars/4 is too generous for agent traffic; measured 3.46 chars/token
   on real history. CHARS_PER_TOKEN 4 -> 3.5 (env-overridable).

Deliberately NOT switching to real tokenization: the bundled tiktoken
tokenizer takes 2.2s on 100k chars of repetitive content and raises
"Max stack size exceeded for backtracking" above ~1M chars — both
unacceptable in a hot path that runs several times per turn over the
whole history. Documented in the module docstring so nobody "upgrades"
it later. ESTIMATOR_OVERHEAD_TOKENS=20k is now validated by the
measurement above (system + tools + framing ~= 22k).

CHARS_PER_TOKEN is a float now; int()-guarded at the three call sites
that use it for slicing.

+++ sandbox_agent/compaction/estimator.py
@@ -44,7 +44,7 @@ def select_tier(messages: List[Message], context_tokens: int = None) -> Tuple[Ti
 
     # Check if truncating tool results alone would be enough
     reducible_chars = _estimate_tool_result_reduction(messages)
-    overflow_chars = overflow * CHARS_PER_TOKEN
+    overflow_chars = int(overflow * CHARS_PER_TOKEN)
     # Need 50% headroom beyond the overflow to be confident truncation alone works
     truncate_threshold = int(overflow_chars * 1.5)
